修改了Item, ItemImage, category三个实体类的一些类型错误

+++ backend/src/main/java/com/yatidle/backend/entity/Item.java
@@ -12,17 +12,17 @@ import java.time.LocalDateTime;
 @TableName("item")
 public class Item {
     @TableId(type = IdType.AUTO)
-    private long id;
-    private long userId;
+    private Long id;
+    private Long userId;
     private String title;
     private String campus;
     private String conditionLevel;
     private String description;
     private BigDecimal price;
-    private long categoryId;
+    private Long categoryId;
     private String status;
-    private long viewCount;
-    private long favoriteCount;
+    private Integer viewCount;
+    private Integer favoriteCount;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
     private Integer isDeleted;
